chore: fix web lint errors
